refactor(messages): enclose configuration in an object factory, improve coroutines dispatcher and dry logic
